+++ movie-service/src/main/java/com/movie_service/repository/TheatreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.movie_service.entity.Theatre;
 
-public interface TheatreRepository extends JpaRepository<Theatre, Integer> {
+public interface TheatreRepository extends JpaRepository<Theatre, Long> {
 
 	List<Theatre> findByName(String name);
 	boolean existsByName(String name);
